Correcting the distribution of H3.3 and K27Ac.
Now, H3.3 happens with a likelyhood defined by user in open chromatin and the rest of the genome gets what's left to make the total to 10%. Also, half of the H3.3 will get K27Ac. This happens separately for each individual in the population so makes the total look like a distribution.